requirements.txt
tweepy==4.14.0
google-api-python-client==2.118.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
python-dotenv==1.0.1 